+++ src/main/resources/static/js/basic.js
@@ -170,93 +170,6 @@ function addProduct(itemDto) {
   });
 }
 
-function showProduct(folderId = null) {
-  /**
-   * 관심상품 목록: #product-container
-   * 검색결과 목록: #search-result-box
-   * 관심상품 HTML 만드는 함수: addProductItem
-   */
-
-  let dataSource = null;
-
-  var sorting = $("#sorting option:selected").val();
-  var isAsc = $(':radio[name="isAsc"]:checked').val();
-
-  if (folderId) {
-    dataSource = `/api/folders/${folderId}/products?sortBy=${sorting}&isAsc=${isAsc}`;
-  } else if(folderTargetId === undefined) {
-    dataSource = `/api/products?sortBy=${sorting}&isAsc=${isAsc}&folderId=${folderId}`;
-  } else {
-    dataSource = `/api/folders/${folderTargetId}/products?sortBy=${sorting}&isAsc=${isAsc}`;
-  }
-
-  $('#product-container').empty();
-  $('#search-result-box').empty();
-  $('#pagination').pagination({
-    dataSource,
-    locator: 'content',
-    alias: {
-      pageNumber: 'page',
-      pageSize: 'size'
-    },
-    totalNumberLocator: (response) => {
-      return response.totalElements;
-    },
-    pageSize: 10,
-    showPrevious: true,
-    showNext: true,
-    ajax: {
-      beforeSend: function () {
-        $('#product-container').html('상품 불러오는 중...');
-      },
-      error(error, status, request) {
-        if (error.status === 403) {
-          $('html').html(error.responseText);
-          return;
-        }
-        logout();
-      }
-    },
-    callback: function (response, pagination) {
-      $('#product-container').empty();
-      for (let i = 0; i < response.length; i++) {
-        let product = response[i];
-        let tempHtml = addProductItem(product);
-        $('#product-container').append(tempHtml);
-      }
-    }
-  });
-}
-
-// Folder 관련 기능
-function openFolder(folderId) {
-  folderTargetId = folderId;
-  $("button.product-folder").removeClass("folder-active");
-  if (!folderId) {
-    $("button#folder-all").addClass('folder-active');
-  } else {
-    $(`button[value='${folderId}']`).addClass('folder-active');
-  }
-  showProduct(folderId);
-}
-
-// 폴더 추가 팝업
-function openAddFolderPopup() {
-  $('#container2').addClass('active');
-}
-
-// 폴더 Input 추가
-function addFolderInput() {
-  $('#folders-input').append(
-      `<input type="text" class="folderToAdd" placeholder="추가할 폴더명">
-       <span onclick="closeFolderInput(this)" style="margin-right:5px">
-            <svg xmlns="http://www.w3.org/2000/svg" width="30px" fill="red" class="bi bi-x-circle-fill" viewBox="0 0 16 16">
-              <path d="M16 8A8 8 0 1 1 0 8a8 8 0 0 1 16 0zM5.354 4.646a.5.5 0 1 0-.708.708L7.293 8l-2.647 2.646a.5.5 0 0 0 .708.708L8 8.707l2.646 2.647a.5.5 0 0 0 .708-.708L8.707 8l2.647-2.646a.5.5 0 0 0-.708-.708L8 7.293 5.354 4.646z"/>
-            </svg>
-       </span>
-      `
-  );
-}
 
 function closeFolderInput(folder) {
   $(folder).prev().remove();
