+++ src/main/resources/templates/additional-info.html
@@ -140,7 +140,7 @@
 
     // Ajax로 서버에 전달
     var xhr = new XMLHttpRequest();
-    xhr.open("POST", "/user/additional-info", true);
+    xhr.open("POST", "/users/additional-info", true);
     xhr.setRequestHeader("Content-Type", "application/json;charset=UTF-8");
     xhr.setRequestHeader("Authorization", "Bearer " + getCookie("Authorization"));
