+++ src/main/resources/templates/login.html
@@ -16,10 +16,10 @@
 <body>
 <div id="login-form">
   <div id="login-title">Login</div>
-  <button id="login-kakao-btn" onclick="location.href='https://kauth.kakao.com/oauth/authorize?client_id=da69a22183d394b6687ef78454a12c34&redirect_uri=http://localhost:8080/api/user/kakao/callback&response_type=code'">
+  <button id="login-kakao-btn" onclick="location.href='https://kauth.kakao.com/oauth/authorize?client_id=da69a22183d394b6687ef78454a12c34&redirect_uri=http://localhost:8080/users/kakao/callback&response_type=code'">
     카카오로 로그인하기
   </button>
-  <button id="login-id-btn" onclick="location.href='/user/signup'">
+  <button id="login-id-btn" onclick="location.href='/users/signup'">
     회원 가입하기
   </button>
   <div>
@@ -72,7 +72,7 @@
     })
     .fail(function (jqXHR, textStatus) {
       alert("Login Fail");
-      window.location.href = host + '/api/user/login-page?error'
+     // window.location.href = host + '/users/login-page?error'
     });
   }
 </script>
