+++ src/main/resources/templates/signup.html
@@ -26,7 +26,7 @@
 <div id="login-form">
   <div id="login-title">Sign up</div>
 
-  <form action="/user/signup" method="post">
+  <form action="/users/signup" method="post">
     <div class="login-id-label">Username</div>
     <input type="text" name="username" placeholder="Username" class="login-input-box">
